Added actions
Awesome list - thank you for putting it together!

I have added some actions we have built for our tooling - hopefully somebody else may find them useful. I know the Dockerlint is a duplicate in functionality but it uses a different tool (could not find any guidance on that in the contributions readme). I am happy to remove it.