Added a guide for deploying Next.js project with Up
Hello! I've made a tiny guide on how to deploy and continuously deliver Next.js project with Up with short introduction to GitHub Actions. Examples and code snippets are written using recently updated Workflow YAML syntax. 

It would be awesome if we add it to Deployment section of the list!